+++ src/pytest_params/request_params.py
@@ -1,7 +1,9 @@
-from typing import Any
+from typing import Optional, TypeVar
+
+T = TypeVar('T')
 
 
-def get_request_param(request, param: str, default: Any = None) -> Any:
+def get_request_param(request, param: str, default: Optional[T] = None) -> Optional[T]:
     if hasattr(request, 'param') and isinstance(request.param, dict):
         return request.param.get(param, default)
     return default
