+++ src/pytest_params/request_params.pyi
@@ -1,5 +1,5 @@
-from typing import Any, TypeVar, Optional
+from typing import Any, Optional, TypeVar
 
 T = TypeVar('T')
 
-def get_request_param(request: Any, param: str, default: T = None) -> Optional[T]: ...
+def get_request_param(request: Any, param: str, default: Optional[T] = None) -> Optional[T]: ...
